LLMApp-Chatbot/requirements.txt
langchain
pinecone-client
openai
tiktoken
nest_asyncio
chromadb
dotenv